Add Force Shield
Press Q to activate a force shield that drains 5 force meter per second. You need to have a minimum of 50 force in order to activate it, and it will remain active as long as you hold Q or have more than 0 force. While the shield is active, any damage will be applied to your force meter instead.